requirements.txt
Faker~=17.3.0
requests~=2.28.1
behave~=1.2.6
selenium~=4.8.2
Appium-Python-Client~=2.8.1
pytest~=7.2.1
python-dateutil~=2.8.2
locust~=2.14.2